requirements.txt
requests==2.26.0
nose==1.3.7
unittest2==1.1.0
mock==4.0.3
PyYAML==6.0